separate platform specific locate
